Added map navigation and logout functionality to my app.
