+++ .idea/misc.xml
@@ -19,6 +19,7 @@
         <entry key="app/src/main/res/layout/activity_phrases.xml" value="0.21113128662109376" />
         <entry key="app/src/main/res/layout/activity_sign_up.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_favorite_phrases.xml" value="0.11005434782608696" />
+        <entry key="app/src/main/res/layout/fragment_map.xml" value="0.1734375" />
         <entry key="app/src/main/res/layout/item_phrase.xml" value="0.3135618348052536" />
         <entry key="app/src/main/res/layout/nav_header_layout.xml" value="0.29179941813151045" />
         <entry key="app/src/main/res/layout/section_header.xml" value="0.15353260869565216" />
