+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.alana.wheretonext" >
+    package="com.alana.wheretonext">
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
@@ -16,19 +16,19 @@
         android:roundIcon="@mipmap/ic_launcher_wtn_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.WhereToNext"
-        tools:targetApi="31" >
+        tools:targetApi="31">
         <activity
-            android:name=".ui.phrases.PhrasesActivity"
+            android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.map.MapActivity"
+            android:name=".ui.phrases.PhrasesActivity"
             android:exported="false" />
         <activity
             android:name=".ui.signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".ui.login.LoginActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
