--- app/src/main/res/layout/activity_favorite_phrases.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_margin="16dp"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/tvFavoritePhraseTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="30dp"
-        android:textStyle="bold"
-        android:text="@string/favorite_phrases" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvFavePhrases"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/tvFavoritePhraseTitle"
-        android:layout_marginTop="10dp"
-        android:padding="8dp" />
-
-</RelativeLayout>
