+++ app/src/main/res/layout/activity_phrases.xml
@@ -1,59 +1,97 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.sothree.slidinguppanel.SlidingUpPanelLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:sothree="http://schemas.android.com/apk/res-auto"
-    android:layout_margin="16dp"
-    android:orientation="vertical"
-    android:gravity="bottom"
-    android:id="@+id/slidingUp"
-    sothree:umanoOverlay="true">
+    android:id="@+id/phrasesDrawerLayout">
 
-    <RelativeLayout
+    <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rvPhrases"
+        <com.google.android.material.appbar.AppBarLayout
+            android:id="@+id/phrasesAppBarLayout"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="40dp"
-            android:padding="8dp" />
+            android:layout_height="wrap_content">
 
-        <TextView
-            android:id="@+id/tvCountryName"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:textColor="@color/navy_blue"
-            tools:text="United States" />
-    </RelativeLayout>
+            <androidx.appcompat.widget.Toolbar
+                android:id="@+id/phrasesToolbar"
+                android:layout_width="match_parent"
+                android:layout_height="?attr/actionBarSize"
+                android:theme="@style/Theme.WhereToNext"
+                app:layout_anchor="@+id/appBarLayout"
+                app:layout_anchorGravity="center" />
 
-    <RelativeLayout
-        android:layout_width="match_parent"
+        </com.google.android.material.appbar.AppBarLayout>
+
+        <com.sothree.slidinguppanel.SlidingUpPanelLayout
+            android:id="@+id/slidingUp"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_margin="16dp"
+            android:gravity="bottom"
+            android:orientation="vertical"
+            app:umanoOverlay="true">
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rvPhrases"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_below="@+id/tvCountryName"
+                    android:layout_marginTop="10dp"
+                    android:padding="8dp" />
+
+                <TextView
+                    android:id="@+id/tvCountryName"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textColor="@color/navy_blue"
+                    android:textSize="30dp"
+                    android:textStyle="bold"
+                    android:layout_marginTop="50dp"
+                    tools:text="United States" />
+            </RelativeLayout>
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="700dp"
+                android:background="@drawable/drawable_round_background">
+
+                <TextView
+                    android:id="@+id/tvFavoritePhraseTitle"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_margin="10dp"
+                    android:text="@string/favorite_phrases"
+                    android:textColor="@color/navy_blue"
+                    android:textSize="30dp"
+                    android:textStyle="bold" />
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rvFavePhrases"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_below="@+id/tvFavoritePhraseTitle"
+                    android:layout_marginTop="10dp"
+                    android:padding="8dp" />
+
+            </RelativeLayout>
+
+        </com.sothree.slidinguppanel.SlidingUpPanelLayout>
+
+    </androidx.coordinatorlayout.widget.CoordinatorLayout>
+
+    <com.google.android.material.navigation.NavigationView
+        android:id="@+id/phrasesNavView"
+        android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:background="@drawable/drawable_round_background">
-
-        <TextView
-            android:id="@+id/tvFavoritePhraseTitle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:text="@string/favorite_phrases"
-            android:textColor="@color/navy_blue"
-            android:layout_margin="10dp" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rvFavePhrases"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/tvFavoritePhraseTitle"
-            android:layout_marginTop="10dp"
-            android:padding="8dp" />
-
-    </RelativeLayout>
-
-</com.sothree.slidinguppanel.SlidingUpPanelLayout>
+        app:menu="@menu/nav_menu"
+        android:layout_gravity="start"
+        app:headerLayout="@layout/nav_header_layout"
+        />
+</androidx.drawerlayout.widget.DrawerLayout>
